requirements.txt
asgiref==3.9.1
asn1crypto==1.5.1
attrs==25.3.0
cbor2==5.7.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
cloudinary==1.44.1
cryptography==45.0.6
defusedxml==0.7.1
Django==5.2.5
django-cors-headers==4.7.0
django-environ==0.12.0
django-extensions==4.1
django-filter==25.1
django-silk==5.4.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
djoser==2.3.3
drf-spectacular==0.28.0
gprof2dot==2025.4.14
idna==3.10
inflection==0.5.1
isort==6.0.1
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
oauthlib==3.3.1
psycopg==3.2.9
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==25.1.0
python3-openid==3.2.0
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.5
requests-oauthlib==2.0.0
rpds-py==0.27.0
six==1.17.0
social-auth-app-django==5.5.1
social-auth-core==4.7.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
